Attributes

These are used to get and set properties of the model.

MathProgBase.AbstractAttribute — Type.

AbstractAttribute

Abstract supertype for attribute objects that can be used to set or get attributes (properties) of the model.

source

MathProgBase.cangetattribute — Function.

cangetattribute(m::AbstractMathProgModel, attr::AbstractAttribute)::Bool

Return a Bool indicating whether the model m currently has a value for the attributed specified by attribute type attr.

source

MathProgBase.getattribute — Function.

getattribute(m::AbstractMathProgModel, attr::AbstractAttribute, extra_args...)

Return an attribute of the model m specified by attribute type attr.

Examples

getattribute(m, ObjectiveValue())
getattribute(m, VariableResult(), ref)
getattribute(m, VariableResult(5), [ref1,ref2])
getattribute(m, OtherAttribute("something specific to cplex"))

source

MathProgBase.cansetattribute — Function.

cansetattribute(m::AbstractMathProgModel, attr::AbstractAttribute)::Bool

Return a Bool indicating whether the model m will accept a setattribute! call for the attributed specified by attribute type attr.

source

MathProgBase.setattribute! — Function.

setattribute!(m::AbstractMathProgModel, attr::AbstractAttribute, ...)

Set an attribute of the model m specified by attribute type attr.

source

Scalar Attributes

MathProgBase.ObjectiveValue — Type.

ObjectiveValue(resultidx::Int=1, objectiveindex::Int=1)

The objective value of the resultindex'th primal result of the objectiveindex'th objective.

Both resultindex and objectiveindex default to 1.

source

MathProgBase.ObjectiveBound — Type.

ObjectiveBound()

The best known bound on the optimal objective value.

source

MathProgBase.RelativeGap — Type.

RelativeGap()

The final relative optimality gap as optimization terminated. That is, $\frac{|b-f|}{|f|}$, where $b$ is the best bound and $f$ is the best feasible objective value.

source

MathProgBase.SolveTime — Type.

SolveTime()

The total elapsed solution time (in seconds) as reported by the solver.

source

MathProgBase.Sense — Type.

Sense()

The optimization sense of the model, an OptimizationSense with value MinSense or MaxSense.

source

MathProgBase.SimplexIterations — Type.

SimplexIterations()

The cumulative number of simplex iterations during the optimization process. In particular, for a MIP the total simplex iterations for all nodes.

source

MathProgBase.BarrierIterations — Type.

BarrierIterations()

The cumulative number of barrier iterations during the optimization process.

source

MathProgBase.NodeCount — Type.

NodeCount()

The total number of branch-and-bound nodes explored.

source

MathProgBase.RawSolver — Type.

RawSolver()

An object that may be used to access a solver-specific API for this model.

source

MathProgBase.ResultCount — Type.

ResultCount()

The number of results available.

source

MathProgBase.VariableCount — Type.

VariableCount()

The number of variables in the model.

source

MathProgBase.ConstraintCount — Type.

ConstraintCount{T}()

The number of constraints of type T in the model.

source

MathProgBase.SupportsVariablesInSet — Type.

SupportsVariablesInSet{T}()

A Bool indicating whether the solver or model supports a constraint of type $x_{varidx} \in S$ where $S$ is a set of type T and $varidx$ indicates any subset of the variables.

source

MathProgBase.SupportsAffineInSet — Type.

SupportsAffineInSet{T}()

A Bool indicating whether the solver or model supports a constraint of of the form "affine expression" in $S$ where $S$ is a set of type T.

source

MathProgBase.SupportsQuadraticInSet — Type.

SupportsQuadraticInSet{T}()

A Bool indicating whether the solver or model supports a constraint of of the form "quadratic expression" in $S$ where $S$ is a set of type T.

source

Variable Attributes

These attributes are associated with variables. Calls to getattribute and setattribute! should include as an argument a single VariableReference or a vector of VariableReference objects.

MathProgBase.VariablePrimalStart — Type.

VariablePrimalStart()

An initial assignment of the variables that the solver may use to warm-start the solve.

source

MathProgBase.VariableLowerBoundDualStart — Type.

VariableLowerBoundDualStart()

An initial assignment of the variable lower-bound duals that the solver may use to warm-start the solve.

source

MathProgBase.VariableUpperBoundDualStart — Type.

VariableUpperBoundDualStart()

An initial assignment of the variable upper-bound duals that the solver may use to warm-start the solve.

source

MathProgBase.VariableLowerBound — Type.

VariableLowerBound()

Lower-bound constraints on variables. -Inf is valid as no bound.

source

MathProgBase.VariableUpperBound — Type.

VariableUpperBound()

Upper-bound constraints for the variables. Inf is valid as no bound.

source

MathProgBase.VariablePrimal — Type.

VariablePrimal(N)
VariablePrimal()

The assignment to the primal variables in result N. If N is omitted, it is 1 by default.

source

MathProgBase.VariableLowerBoundDual — Type.

VariableLowerBoundDual(N)

The assignment to the duals on the variable lower bounds in result N. If N is omitted, it is interpreted as 1.

source

MathProgBase.VariableUpperBoundDual — Type.

VariableUpperBoundDual(N)

The assignment to the duals on the variable upper bounds in result N. If N is omitted, it is interpreted as 1.

source

Constraint Attributes

These attributes are associated with constraints. Calls to getattribute and setattribute! should include as an argument a single ConstraintReference or a vector of ConstriaintReference{T} objects.

MathProgBase.ConstraintPrimalStart — Type.

ConstraintPrimalStart()

An initial assignment of the constraint primal values that the solver may use to warm-start the solve.

source

MathProgBase.ConstraintDualStart — Type.

ConstraintDualStart()

An initial assignment of the constriant duals that the solver may use to warm-start the solve.

source

MathProgBase.ConstraintPrimal — Type.

ConstraintPrimal(N)
ConstraintPrimal()

The assignment to the constraint primal values in result N. If N is omitted, it is 1 by default.

source

MathProgBase.ConstraintDual — Type.

ConstraintDual(N)
ConstraintDual()

The assignment to the constraint dual values in result N. If N is omitted, it is 1 by default.

source