Basics

Some introduction to MPB API. List basic standalone methods.

MathProgBase.AbstractModel — Type.

AbstractModel

Abstract supertype which represents a solver's in-memory representation of an optimization problem.

source

MathProgBase.AbstractNLPModel — Type.

AbstractNLPModel

Abstract supertype which represents a solver's in-memory representation of a non-linear optimization problem.

source

MathProgBase.AbstractMathProgSolver — Type.

AbstractMathProgSolver

Abstract supertype for "solver" objects. A solver is a lightweight object used for selecting solvers and parameters. It does not store any instance data.

source

MathProgBase.Model — Function.

Model(solver::AbstractMathProgSolver)

Create an instance of AbstractModel using the given solver.

source

MathProgBase.NLPModel — Function.

NLPModel(solver::AbstractMathProgSolver)

Create an instance of AbstractNLPModel using the given solver.

source

MathProgBase.optimize! — Function.

optimize!(m::AbstractMathProgModel)

Start the solution procedure.

source

MathProgBase.freemodel! — Function.

freemodel!(m::AbstractMathProgModel)

Release any resources and memory used by the model. Note that the Julia garbage collector takes care of this automatically, but automatic collection cannot always be forced. This method is useful for more precise control of resources, especially in the case of commercial solvers with licensing restrictions on the number of concurrent runs. Users must discard the model object after this method is invoked.

source