High-level Interfaces

High-level Interfaces

These docs are out of date. We need to update the high-level interfaces for new statuses.

Linear Programming

MathProgBase.linprog — Function.

linprog(c::InputVector, A::AbstractMatrix, rowlb::InputVector, rowub::InputVector, lb::InputVector, ub::InputVector, solver::AbstractMathProgSolver)

This function allows one to specify two-sided linear constraints (also known as range constraints) to solve the linear programming problem:

\[\begin{align*} \text{min}_{x} \quad & c^{T} x \\ \text{s.t.} \quad & rowlb \leq A^{T} x \leq rowub \\ \quad & l \leq x \leq u \\ \end{align*}\]

where:

$c$ is the objective vector, always in the sense of minimization

$A$ is the constraint matrix

$rowlb$ is the vector of row lower bounds

$rowub$ is the vector of row upper bounds

$lb$ is the vector of lower bounds on the variables

$ub$ is the vector of upper bounds on the variables, and

$solver$ specifies the desired solver, see Choosing Solvers.

A scalar is accepted for the $l$, $u$, $rowlb$, and $rowub$ arguments, in which case its value is replicated. The values $-Inf$ and $Inf$ are interpreted to mean that there is no corresponding lower or upper bound. Equality constraints are specified by setting the row lower and upper bounds to the same value.

A variant usage of this function is to consider the linear programming problem in the following form,

linprog(c::InputVector, A::AbstractMatrix, sense::InputVector, b::InputVector, lb::InputVector, ub::InputVector, solver::AbstractMathProgSolver)

\[\begin{align*} \text{min}_{x} \quad & c^{T}x \\ \text{s.t.} \quad & A_{i}^{T} \, \text{sense}_{i} \, b_{i} \quad \forall i \\ \quad & l \leq x \leq u \end{align*}\]

where:

$c$: is the objective vector, always in the sense of minimization

$A$: is the constraint matrix, with rows a_i (viewed as column-oriented vectors)

sense: is a vector of constraint sense characters $<$, $=$, and $>$

$b$: is the right-hand side vector

$l$: is the vector of lower bounds on the variables

$u$ is the vector of upper bounds on the variables, and solver specifies the desired solver, see Choosing Solvers.

A shortened version is defined as::

linprog(c, A, lb, ub, solver) = linprog(c, A, lb, ub, 0, Inf, solver)

Note

The function linprog calls two independent functions for building and solving the linear programming problem, namely buildlp and solvelp.

The linprog function returns an instance of the type::

type LinprogSolution
    status
    objval
    sol
    attrs
end

where status is a termination status symbol, one of :Optimal, :Infeasible, :Unbounded, :UserLimit (iteration limit or timeout), :Error (and maybe others).

If status is :Optimal, the other members have the following values:

objval – optimal objective value

sol – primal solution vector

attrs – a dictionary that may contain other relevant attributes such as:

redcost – dual multipliers for active variable bounds (zero if inactive)

lambda – dual multipliers for active linear constraints (equalities are always active)

If status is :Infeasible, the attrs member will contain an infeasibilityray if available; similarly for :Unbounded problems, attrs will contain an unboundedray if available.

colbasis – optimal simplex basis statuses for the variables (columns) if available. Possible values are :NonbasicAtLower, :NonbasicAtUpper, :Basic, and :Superbasic (not yet implemented by any solvers)

rowbasis – optimal simplex basis statuses for the constraints (rows) if available (not yet implemented by any solvers)

For example, we can solve the two-dimensional problem (see test/linprog.jl):

\[ \begin{align*} \text{min}_{x,y} \quad & -x \\ \text{s.t.} \quad & 2x + y \leq 1.5 \\ \quad & x \geq 0, y \geq 0 \end{align*}\]

using MathProgBase, Clp

sol = linprog([-1,0],[2 1],'<',1.5, ClpSolver())
if sol.status == :Optimal
    println("Optimal objective value is $(sol.objval)")
    println("Optimal solution vector is: [$(sol.sol[1]), $(sol.sol[2])]")
else
    println("Error: solution status $(sol.status)")
end

source

MathProgBase.buildlp — Function.

buildlp(c::InputVector, A::AbstractMatrix, sense::InputVector, b::InputVector, lb::InputVector, ub::InputVector, solver::AbstractMathProgSolver)

Builds the linear programming problem as defined in linprog and accepts the following arguments:

$c$ is the objective vector, always in the sense of minimization

$A$ is the constraint matrix

$sense$ is a vector of constraint sense characters $'<'$, $'='$, and $'>'$

$b$ is the right-hand side vector

$l$ is the vector of lower bounds on the variables

$u$ is the vector of upper bounds on the variables, and

$solver$ specifies the desired solver, see Choosing Solvers.

A scalar is accepted for the $b$, $sense$, $l$, and $u$ arguments, in which case its value is replicated. The values $-Inf$ and $Inf$ are interpreted to mean that there is no corresponding lower or upper bound.

This variant of buildlp allows to specify two-sided linear constraints (also known as range constraints) similar to linprog, and accepts the following arguments:

buildlp(c::InputVector, A::AbstractMatrix, rowlb::InputVector, rowub::InputVector, lb::InputVector, ub::InputVector, solver::AbstractMathProgSolver)

$c$ is the objective vector, always in the sense of minimization

$A$ is the constraint matrix

$rowlb$ is the vector of row lower bounds

$rowub$ is the vector of row upper bounds

$lb$ is the vector of lower bounds on the variables

$ub$ is the vector of upper bounds on the variables, and

$solver$ specifies the desired solver, see Choosing Solvers.

A scalar is accepted for the $l$, $u$, $lb$, and $ub$ arguments, in which case its value is replicated. The values $-Inf$ and $Inf$ are interpreted to mean that there is no corresponding lower or upper bound. Equality constraints are specified by setting the row lower and upper bounds to the same value.

The buildlp function returns an AbstractLinearQuadraticModel that can be input to solvelp in order to obtain a solution.

source

MathProgBase.solvelp — Function.

Solves the linear programming problem as defined in linprog and accepts the following argument:

m is an AbstractLinearQuadraticModel (e.g., as returned by buildlp).

The solvelp function returns an instance of the type::

type LinprogSolution
    status
    objval
    sol
    attrs
end

source

Mixed-integer Programming

MathProgBase.mixintprog — Function.

mixintprog(c::InputVector, A::AbstractMatrix, sense::InputVector, b::InputVector, vartypes::SymbolInputVector, lb::InputVector, ub::InputVector, solver::AbstractMathProgSolver)

Solves the same optimization problem as linprog above, except variables are additionally constrained to take only integer values if the corresponding entry in the varypes vector is the symbol :Int. Continuous variables are indicated by the value :Cont, binary variables should be specified by :Bin, semicontinuous by :SemiCont, and semi-integer by :SemiInt.

A scalar is accepted for the $sense$, $b$, $vartypes$, $lb$, and $ub$ arguments, in which case its value is replicated. The values $-Inf$ and $Inf$ are interpreted to mean that there is no corresponding lower or upper bound.

The mixintprog function returns an instance of the type::

type MixintprogSolution
    status
    objval
    sol
    attrs
end

where status takes the same values as with linprog.

If status does not indicate error or infeasiblity, the other members have the following values:

objval – optimal objective value

sol – primal solution vector

attrs – a dictionary that may contain other relevant attributes such as:

objbound – Best known lower bound on the objective value

Analogous shortened and range-constraint versions are available as well.

We can solve a binary knapsack problem

\[\begin{align*} \text{max} \, & 5x_1 + 3x_2 + 2x_3 + 7x_4 + 4x_5 \\ \text{s.t.} & 2x_1 + 8x_2 + 4x_3 + 2x_4 + 5x_5 \leq 10 \\ & (x_1, x_2, x_3, x_4, x_5) \in \{0,1\}^5 \end{align*}\]

with the following code

mixintprog(-[5.,3.,2.,7.,4.],[2. 8. 4. 2. 5.],'<',10,:Int,0,1,CbcSolver())

source

Quadratic Programming

MathProgBase.quadprog — Function.

quadprog(c::InputVector, Q::AbstractMatrix, A::AbstractMatrix, sense::InputVector, b::InputVector, lb::InputVector, ub::InputVector, solver::AbstractMathProgSolver)

Solves the quadratic programming problem:

\[ \begin{align*} \text{min}_{x} \quad & \frac{1}{2}x^TQx + c^Tx \\ \text{s.t.} \quad & a_i^Tx \, \text{sense}_i \, b_i \forall \quad i \\ \quad & l \leq x \leq u \\ \end{align*}\]

where:

$c$ is the objective vector, always in the sense of minimization

$Q$ is the Hessian matrix of the objective

$A$ is the constraint matrix, with rows :math:a_i (viewed as column-oriented vectors)

$sense$ is a vector of constraint sense characters $<$, $=$, and $>$

$b$ is the right-hand side vector

$l$ is the vector of lower bounds on the variables

$u$ is the vector of upper bounds on the variables, and

$solver$ specifies the desired solver, see Choosing Solvers.

A scalar is accepted for the $b$, $sense$, $l$, and $u$ arguments, in which case its value is replicated. The values $-Inf$ and $Inf$ are interpreted to mean that there is no corresponding lower or upper bound.

.. note:: Quadratic programming solvers extensively exploit the sparsity of the Hessian matrix $Q$ and the constraint matrix $A$. While both dense and sparse matrices are accepted, for large-scale problems sparse matrices should be provided if permitted by the problem structure.

The quadprog function returns an instance of the type::

type QuadprogSolution
    status
    objval
    sol
    attrs
end

where status is a termination status symbol, one of :Optimal, :Infeasible, :Unbounded, :UserLimit (iteration limit or timeout), :Error (and maybe others).

If status is :Optimal, the other members have the following values:

objval – optimal objective value

sol – primal solution vector

attrs – a dictionary that may contain other relevant attributes (not currently used).

Analogous shortened and range-constraint versions are available as well.

We can solve the three-dimensional QP (see test/quadprog.jl):

\[ \begin{align*} \text{min}_{x,y,z} \quad & x^2+y^2+z^2+xy+yz \\ \text{s.t.} \quad & x + 2y + 3z \geq 4 \\ \quad & x + y \geq 1 \end{align*}\]

using MathProgBase, Ipopt

sol = quadprog([0., 0., 0.],[2. 1. 0.; 1. 2. 1.; 0. 1. 2.],[1. 2. 3.; 1. 1. 0.],'>',[4., 1.],-Inf,Inf, IpoptSolver())
if sol.status == :Optimal
    println("Optimal objective value is $(sol.objval)")
    println("Optimal solution vector is: [$(sol.sol[1]), $(sol.sol[2]), $(sol.sol[3])]")
else
    println("Error: solution status $(sol.status)")
end

source