Sets

List of sets.

MathProgBase.NonNegative — Type.

NonNegative(n)

The nonnegative orthant $\{ x \in \mathbb{R}^n : x \ge 0 \}$ where the dimension $n$ is specified by the field n.

source

MathProgBase.NonPositive — Type.

NonPositive(n)

The nonpositive orthant $\{ x \in \mathbb{R}^n : x \le 0 \}$ where the dimension $n$ is specified by the field n.

source

MathProgBase.Zero — Type.

Zero(n)

The set $\{0\}^n$ where the dimension $n$ is specified by the field n.

source

MathProgBase.Interval — Type.

Interval(lower,upper)

The set $[l,u] \subseteq \mathbb{R}^n$ where $l$ and $u$ are specified by lower and upper, respectively. We allow lower and upper to be -Inf or Inf, in which case the set is interpreted as a one-sided interval.

source

MathProgBase.SecondOrderCone — Type.

SecondOrderCone(n)

The second-order cone or the Lorenz cone of dimension n defined as

\[\{ (t,x) \in \mathbb{R}^n : t \ge ||x||_2 \}.\]

source

MathProgBase.Integers — Type.

Integers(n)

The set of integers $\mathbb{Z}^n$.

source

MathProgBase.Binaries — Type.

Binaries(n)

The set of binary vectors $\{0,1\}^n$.

source

MathProgBase.SOS1 — Type.

SOS1(weights::Vector{T}) where T

The set corresponding to the special ordered set (SOS) constraint of type 1. Of the variables in the set, at most one can be nonzero. The $weights$ induce an ordering of the variables; as such, they should be unique values. The $k$-th element in the set corresponds to the $k$-th weight in $weights$.

See here for a description of SOS constraints and their potential uses.

source

MathProgBase.SOS2 — Type.

SOS2(weights::Vector{T}) where T

The set corresponding to the special ordered set (SOS) constraint of type 2. Of the variables in the set, at most two can be nonzero, and if two are nonzero, they must be adjacent in the ordering of the set. The $weights$ induce an ordering of the variables; as such, they should be unique values. The $k$-th element in the set corresponds to the $k$-th weight in $weights$.

See here for a description of SOS constraints and their potential uses.

source