Status Codes

Termination Status

The TerminationStatus attribute is meant to explain the reason why the solver stopped executing. The value of the attribute is of type TerminationStatusCode.

MathProgBase.TerminationStatus — Type.

TerminationStatus()

A TerminationStatusCode explaining why the solver stopped.

source

MathProgBase.TerminationStatusCode — Type.

TerminationStatusCode

An Enum of possible values for the TerminationStatus attribute. This attribute is meant to explain the reason why the solver stopped executing.

OK

These are generally OK statuses.

Success: the algorithm ran successfully and has a result. This includes cases where the algorithm converges to an infeasible point (NLP) or converges to a solution of a homogeneous self-dual problem and has a certificate of primal/dual infeasibility.

AlmostSuccess: the algorithm almost ran successfully (e.g., to relaxed convergence tolerances) and has a result.

InfeasibleNoResult: the algorithm stopped because it decided that the problem is infeasible but does not have a result to return.

UnboundedNoResult: the algorithm stopped because it decided that the problem is unbounded but does not have a result to return.

InfeasibleOrUnbounded: the algorithm stopped because it decided that the problem is infeasible or unbounded; no result is available. This occasionally happens during MIP presolve.

Limits

The solver stopped because of some user-defined limit. To be documented: IterationLimit, TimeLimit, NodeLimit, SolutionLimit, MemoryLimit, ObjectiveLimit, NormLimit, OtherLimit.

Problematic

This group of statuses means that something unexpected or problematic happened.

SlowProgress: the algorithm stopped because it was unable to continue making progress towards the solution. AlmostSuccess should be used if there is additional information that relaxed convergence tolerances are satisfied.

To be documented: NumericalError, InvalidModel, InvalidOption, Interrupted, OtherError.

source

Result Status

The PrimalStatus and DualStatus attributes are meant to explain how to interpret the result returned by the solver. The value of the attributes are of type ResultStatusCode.

MathProgBase.PrimalStatus — Type.

PrimalStatus(N)
PrimalStatus()

The ResultStatusCode of the primal result N. If N is omitted, it defaults to 1.

source

MathProgBase.DualStatus — Type.

DualStatus(N)
DualStatus()

The ResultStatusCode of the dual result N. If N is omitted, it defaults to 1.

source

MathProgBase.ResultStatusCode — Type.

ResultStatusCode

An Enum of possible values for the PrimalStatus and DualStatus attributes. The values indicate how to interpret the result vector.

FeasiblePoint

NearlyFeasiblePoint

InfeasiblePoint

InfeasibilityCertificate

NearlyInfeasibilityCertificate

ReductionCertificate

NearlyReductionCertificate

Unknown

Other

source

Basis Status

TODO: attributes and status codes for LP basis status