Variables

MathProgBase.VariableReference — Type.

VariableReference

A lightweight object used to reference variables in a model.

source

MathProgBase.candelete — Method.

candelete(m::AbstractMathProgModel, ref::VariableReference)::Bool

Return a Bool indicating whether this variable can be removed from the model m.

source

MathProgBase.isvalid — Method.

isvalid(m::AbstractMathProgModel, ref::VariableReference)::Bool

Return a Bool indicating whether this reference is valid for an active variable in the model m.

source

Base.delete! — Method.

delete!(m::AbstractMathProgModel, ref::VariableReference)

Delete the referenced variable from the model.

delete!(m::AbstractMathProgModel, refs::Vector{VariableReference})

Delete the referenced variables in the vector refs from the model.

source

MathProgBase.addvariables! — Function.

addvariables!(m::AbstractMathProgModel, N::Int)::Vector{VariableReference}

Add N scalar variables to the model, returning a vector of variable references.

source

MathProgBase.addvariable! — Function.

addvariable!(m::AbstractMathProgModel)::VariableReference

Add a scalar variable to the model, returning a variable reference.

source