Solver Interface API

Solver Interface API

Some introduction to MPB API. List basic standalone methods.

MathProgBase.AbstractModel — Type.

AbstractModel

Abstract supertype which represents a solver's in-memory representation of an optimization problem.

source

MathProgBase.AbstractNLPModel — Type.

AbstractNLPModel

Abstract supertype which represents a solver's in-memory representation of a non-linear optimization problem.

source

MathProgBase.AbstractMathProgSolver — Type.

AbstractMathProgSolver

Abstract supertype for "solver" objects. A solver is a lightweight object used for selecting solvers and parameters. It does not store any instance data.

source

MathProgBase.Model — Function.

Model(solver::AbstractMathProgSolver)

Create an instance of AbstractModel using the given solver.

source

MathProgBase.NLPModel — Function.

NLPModel(solver::AbstractMathProgSolver)

Create an instance of AbstractNLPModel using the given solver.

source

MathProgBase.optimize! — Function.

optimize!(m::AbstractMathProgModel)

Start the solution procedure.

source

MathProgBase.freemodel! — Function.

freemodel!(m::AbstractMathProgModel)

Release any resources and memory used by the model. Note that the Julia garbage collector takes care of this automatically, but automatic collection cannot always be forced. This method is useful for more precise control of resources, especially in the case of commercial solvers with licensing restrictions on the number of concurrent runs. Users must discard the model object after this method is invoked.

source

Variables

MathProgBase.VariableReference — Type.

VariableReference

A lightweight object used to reference variables in a model.

source

MathProgBase.candelete — Method.

candelete(m::AbstractMathProgModel, ref::VariableReference)::Bool

Return a Bool indicating whether this variable can be removed from the model m.

source

MathProgBase.isvalid — Method.

isvalid(m::AbstractMathProgModel, ref::VariableReference)::Bool

Return a Bool indicating whether this reference is valid for an active variable in the model m.

source

Base.delete! — Method.

delete!(m::AbstractMathProgModel, ref::VariableReference)

Delete the referenced variable from the model.

delete!(m::AbstractMathProgModel, refs::Vector{VariableReference})

Delete the referenced variables in the vector refs from the model.

source

MathProgBase.addvariables! — Function.

addvariables!(m::AbstractMathProgModel, N::Int)::Vector{VariableReference}

Add N scalar variables to the model, returning a vector of variable references.

source

MathProgBase.addvariable! — Function.

addvariable!(m::AbstractMathProgModel)::VariableReference

Add a scalar variable to the model, returning a variable reference.

In addition, there is a special case for adding variables to existing linear problems.

addvariable!(m::AbstractMathProgModel,
    cref::Vector{Union{
            AffineConstraintRef{NonPositive},
            AffineConstraintRef{NonNegative},
            AffineConstraintRef{Zero},
            AffineConstraintRef{Interval}
        }},
    coefs)::VariableReference

Add a variable with coefficients specified by coefs in the existing affine constraints given by the constraint references cref. If you want to add a variable with coefficients in a constraint that is not listed here (such as a quadratic term, or in the SOC), use addvariable!(m) and then modifyconstraint! instead.

source

Objectives

How to add and set objectives.

MathProgBase.setobjective! — Function.

setobjective!(m::AbstractMathProgModel, b, a_varref::Vector{VariableReference}, a_coef, Q_vari::Vector{VariableReference}, Q_varj::Vector{VariableReference}, Q_coef, N::Int=1)

Set the N'th objective in the model m to be

\[a^Tx + b + \frac{1}{2}x^TQx\]

where $a$ is a sparse vector specified in tuple form by a_varref, and a_coef; $b$ is a scalar; and the symmetric matrix $Q$ is defined by the triplets in Q_vari, Q_varj, Q_coef.

Duplicate indices in either the $a$ vector or the $Q$ matrix are accepted and will be summed together. Off-diagonal entries of $Q$ will be mirrored, so either the upper triangular or lower triangular entries of $Q$ should be provided. If entries for both $(i,j)$ and $(j,i)$ are provided, these are considered duplicate terms. a_varref, Q_vari, Q_varj should be collections of VariableReference objects.

setobjective!(m::AbstractMathProgModel, b, a_varref::Vector{VariableReference}, a_coef, N::Int=1)

Set the N'th objective in the model m to be

\[a^Tx + b\]

where $a$ is a sparse vector specified in tuple form by a_varref and a_coef and $b$ is a scalar.

Duplicate indices in either the $a$ vector are accepted and will be summed together.

source

MathProgBase.modifyobjective! — Function.

modifyobjective!(m::AbstractMathProgModel, i::Int, args...)

Modify elements of the i'th objective depending on the arguments args. The i'th objective will have the form:

\[ a_i^Tx + b_i + \frac{1}{2}x^TQ_ix\]

There are three cases.

Modify Constant term

modifyobjective!(m::AbstractMathProgModel, i::Int, b)

Set the constant term of the i'th row objective to b.

Examples

modifyobjective!(m, 1, 1.0)

Modify Linear term

modifyobjective!(m::AbstractMathProgModel, i::Int, a_varidx, a_coef)

Set elements given by a_varidx in the linear term of the i'th objective to a_coef. Either a_varidx and a_coef are both singletons, or they should be collections with equal length.

The behaviour of duplicate entries in a_varidx is undefined.

Examples

modifyobjective!(m, 1, v, 1.0)
modifyobjective!(m, 1, [v1, v2], [1.0, 2.0])

Modify Quadratic term

modifyobjective!(m::AbstractMathProgModel, i::Int, Q_vari, Q_varj, Q_coef)

Set the elements in the quadratic term of the i'th objective specified by the triplets Q_vari, Q_varj, and Q_coef. Off-diagonal entries will be mirrored. Q_vari, Q_varj should be collections of VariableReference objects.

The behaviour of duplicate entries is undefined. If entries for both $(i,j)$ and $(j,i)$ are provided, these are considered duplicate terms.

Examples

modifyobjective!(m, 1, v1, v2, 1.0)
modifyobjective!(m, 1, [v1, v2], [v1, v1], [1.0, 2.0])

source

MathProgBase.getobjectiveconstant — Function.

getobjectiveconstant(m, i::Int=1)

Return the constant term in the i'th objective.

source

MathProgBase.getobjectiveaffine — Function.

getobjectiveaffine(m, i::Int=1)

Return the affine part of the i'th objective in tuple form (varref,coef) where varref is a VariableReference, and coef is a coefficient. Output is a tuple of two vectors.

getobjectiveaffine(m, v::VariableReference, i::Int=1)

Return the coefficient for the variable v in the affine part of the i'th objective.

source

Constraints

How to add and modify constraints.

MathProgBase.VariablewiseConstraintReference — Type.

VariablewiseConstraintReference{T}

A lightweight object used to reference variablewise constraints in a model. The parameter T is the type of set constraint referenced.

source

MathProgBase.AffineConstraintReference — Type.

AffineConstraintReference{T}

A lightweight object used to reference affine-in-set constraints in a model. The parameter T is the type of set constraint referenced.

source

MathProgBase.QuadraticConstraintReference — Type.

QuadraticConstraintReference{T}

A lightweight object used to reference quadratic-in-set constraints in a model. The parameter T is the type of set constraint referenced.

source

MathProgBase.candelete — Method.

candelete(m::AbstractMathProgModel, ref::ConstraintReference)::Bool

Return a Bool indicating whether this constraint can be removed from the model m.

source

MathProgBase.isvalid — Method.

isvalid(m::AbstractMathProgModel, ref::ConstraintReference)::Bool

Return a Bool indicating whether this reference is valid for an active constraint in the model m.

source

Base.delete! — Method.

delete!(m::AbstractMathProgModel, ref::ConstraintReference)

Delete the referenced constraint from the model.

delete!(m::AbstractMathProgModel, refs::Vector{ConstraintReference})

Delete the referenced constraints in the vector refs from the model.

source

MathProgBase.addconstraint! — Function.

addconstraint!(m::AbstractMathProgModel, b, a_constridx, a_varref::Vector{VariableReference}, a_coef, Q_constridx, Q_vari::Vector{VariableReference}, Q_varj::Vector{VariableReference}, Q_coef, S::AbstractSet)::QuadraticConstraintReference{typeof(S)}

Add the quadratic-in-set constraint

\[Ax + b + q(x) \in S\]

where $A$ is a sparse matrix specified in triplet form by a_constridx, a_varref, and a_coef; $b$ is a vector; $q(x)$ is a vector with component $(q(x))_k$ defined to be $\frac{1}{2}x^TQ_kx$ where the symmetric matrix $Q_k$ is defined by the triplets in Q_vari, Q_varj, Q_coef for which Q_constridx equals k; and the set $S$ is defined by S.

Duplicate indices in either the $A$ or the $Q$ matrix are accepted and will be summed together. Off-diagonal entries of $Q$ will be mirrored, so either the upper triangular or lower triangular entries of $Q$ should be provided. If entries for both $(i,j)$ and $(j,i)$ are provided, these are considered duplicate terms. a_varref, Q_vari, Q_varj should be collections of VariableReference objects.

addconstraint!(m::AbstractMathProgModel, b, a_varref::Vector{VariableReference}, a_coef, Q_vari::Vector{VariableReference}, Q_varj::Vector{VariableReference}, Q_coef, S::AbstractSet)::QuadraticConstraintReference{typeof(S)}

A specialized version of addconstraint! for one-dimensional sets. Add the constraint

\[a^Tx + b + \frac{1}{2}x^TQx \in S\]

where $a$ is a sparse vector specified in tuple form by a_varref, and a_coef; $b$ is a scalar; the symmetric matrix $Q$ is defined by the triplets in Q_vari, Q_varj, Q_coef; and the set $S$ is defined by S.

addconstraint!(m::AbstractMathProgModel, b, a_constridx, a_varref::Vector{VariableReference}, a_coef, S::AbstractSet)::AffineConstraintReference{typeof(S)}

Add the affine-in-set constraint

\[Ax + b \in S\]

where $A$ is a sparse matrix specified in triplet form by a_constridx, a_varref, and a_coef; $b$ is a vector; and the set $S$ is defined by S.

Duplicate indices either $A$ are accepted and will be summed together.

addconstraint!(m::AbstractMathProgModel, b, a_varref::Vector{VariableReference}, a_coef, S::AbstractSet)::AffineConstraintReference{typeof(S)}

A specialized version of addconstraint! for one-dimensional sets. Add the constraint

\[a^Tx + b \in S\]

where $a$ is a sparse vector specified in tuple form by a_varref, and a_coef; $b$ is a scalar; and the set $S$ is defined by S.

addconstraint!(m::AbstractMathProgModel, varref::Vector{VariableReference}, S::AbstractSet)::VariablewiseConstraintReference{typeof(S)}

A specialized version of addconstraint! for variablewise constraints. Add the constraint

\[x_{varref} \in S\]

where varref is a vector of variable references to specifiy the subset of the subvector of x.

source

MathProgBase.modifyconstraint! — Function.

modifyconstraint!(m::AbstractMathProgModel, c::ConstraintReference, i::Int, args...)

Modify elements of the i'th row of the constraint c depending on the arguments args. The i'th row will have the form

\[ a_i^Tx + b_i + \frac{1}{2}x^TQ_ix \in S\]

There are three cases.

Modify Constant term

modifyconstraint!(m::AbstractMathProgModel, c::ConstraintReference, i::Int, b)

Set the constant term of the i'th row in the constraint c to b.

Examples

modifyconstraint!(m, c, 1, 1.0)

Modify Linear term

modifyconstraint!(m::AbstractMathProgModel, c::ConstraintReference, i::Int, a_varref::Vector{VariableReference}, a_coef)

Set elements given by a_varref in the linear term of the i'th element in the constraint c to a_coef. Either a_varref and a_coef are both singletons, or they should be collections with equal length.

The behaviour of duplicate entries in a_varref is undefined.

Examples

modifyconstraint!(m, c, v, 1.0)
modifyconstraint!(m, c, [v1, v2], [1.0, 2.0])

Modify Quadratic term

modifyconstraint!(m::AbstractMathProgModel, c::ConstraintReference, i::Int, Q_vari, Q_varj, Q_coef)

Set the elements in the quadratic term of the i'th element of the constraint c specified by the triplets Q_vari, Q_varj, and Q_coef. Off-diagonal entries will be mirrored. Q_vari, Q_varj should be collections of VariableReference objects.

The behaviour of duplicate entries is undefined. If entries for both $(i,j)$ and $(j,i)$ are provided, these are considered duplicate terms.

Examples

modifyconstraint!(m, c, v1, v2, 1.0)
modifyconstraint!(m, c, [v1, v2], [v1, v1], [1.0, 2.0])

Modify Set

modifyconstraint!(m::AbstractMathProgModel, c::ConstraintReference{S}, set::S)

Change the set of constraint c to the new set set which should be of the same type as the original set.

Examples

If c is a ConstraintReference{Interval}

modifyconstraint!(m, c, Interval(0, 5))
modifyconstraint!(m, c, NonPositive) # errors

source

MathProgBase.getconstraintconstant — Function.

getconstraintconstant(m::AbstractMathProgModel, c::ConstraintReference, i::Int)

Return the constant term of the ith row of the constraint corresponding to c.

source

MathProgBase.getconstraintaffine — Function.

getconstraintaffine(m::AbstractMathProgModel, c::ConstraintReference)

Return the $A$ matrix of the constraint corresponding to c in triplet form (row,varref,coef) where row is an integer, varref is a VariableReference, and coef is a coefficient. Output is a tuple of three vectors.

getconstraintaffine(m::AbstractMathProgModel, c::ConstraintReference, i::Int)

Return the ith row of the $A$ matrix of the constraint corresponding to c in tuple form (varref,coef) where varref is a VariableReference, and coef is a coefficient. Output is a tuple of two vectors.

getconstraintaffine(m::AbstractMathProgModel, c::ConstraintReference, i::Int, v::VariableReference)

Return the element of the $A$ matrix of the constraint corresponding to c in row i and variable v.

source

MathProgBase.getconstraintquadratic — Function.

getconstraintquadratic(m::AbstractMathProgModel, c::ConstraintReference, i::Int)

Return the $Q$ matrix of the ith row of the constraint corresponding to c in triplet form (varref_a,varref_b,coef) where varref_a is a VariableReference, varref_b is a VariableReference, and coef is a coefficient. Output is a tuple of three vectors. The $Q$ matrix must be symmetric, and only one of the two symmetric elements is returned.

getconstraintquadratic(m::AbstractMathProgModel, c::ConstraintReference, i::Int, v1::VariableReference, v2::VariableReference)

Return the element (v1,v2) of the $Q$ matrix of the ith row of the constraint corresponding to c.

source

Sets

List of sets.

MathProgBase.NonNegative — Type.

NonNegative(dim)

The nonnegative orthant $\{ x \in \mathbb{R}^{dim} : x \ge 0 \}$ of dimension $dim$.

source

MathProgBase.NonPositive — Type.

NonPositive(dim)

The nonpositive orthant $\{ x \in \mathbb{R}^{dim} : x \le 0 \}$ of dimension $dim$.

source

MathProgBase.Zero — Type.

Zero(dim)

The set $\{ 0 \}^{dim}$ (containing only the origin) of dimension $dim$.

source

MathProgBase.Interval — Type.

Interval(lower,upper)

The box $[lower, upper] \subseteq \mathbb{R}^{dim}$ where $lower$ and $upper$ are scalars (1-dimensional) or vectors of dimension $dim$. If $lower$ or $upper$ is all $-Inf$ or $Inf$, the set is interpreted as a one-sided interval.

source

MathProgBase.SecondOrderCone — Type.

SecondOrderCone(dim)

The second-order cone (or Lorenz cone) $\{ (t,x) \in \mathbb{R}^{dim} : t \ge || x ||_2 \}$ of dimension $dim$.

source

MathProgBase.PositiveSemidefiniteConeTriangle — Type.

PositiveSemidefiniteConeTriangle(dim)

The (vectorized) cone of symmetric positive semidefinite matrices, with off-diagonals unscaled. The entries of the upper triangular part of the matrix are given row by row (or equivalently, the entries of the lower triangular part are given column by column). An $n \times n$ matrix has $n(n+1)/2$ lower-triangular elements, so for the vectorized cone of dimension $dim$, the corresponding symmetric matrix has side dimension $\sqrt (1/4 + 2 dim) - 1/2$ elements. The scalar product is the sum of the pairwise product of the diagonal entries plus twice the sum of the pairwise product of the upper diagonal entries.

Examples

The matrix

\[\begin{bmatrix} 1 & 2 & 3\\ 2 & 4 & 5\\ 3 & 5 & 6 \end{bmatrix}\]

corresponds to $(1, 2, 3, 4, 5, 6)$ for PositiveSemidefiniteConeTriangle

source

MathProgBase.PositiveSemidefiniteConeScaled — Type.

PositiveSemidefiniteConeScaled(dim)

The (vectorized) cone of symmetric positive semidefinite matrices, with off-diagonals scaled. The entries of the upper triangular part of the matrix are given row by row (or equivalently, the entries of the lower triangular part are given column by column). An $n \times n$ matrix has $n(n+1)/2$ lower-triangular elements, so for the vectorized cone of dimension $dim$, the corresponding symmetric matrix has side dimension $\sqrt (1/4 + 2 dim) - 1/2$ elements. The off-diagonal entries of the matrices of both the cone and its dual are scaled by $\sqrt{2}$ and the scalar product is simply the sum of the pairwise product of the entries.

Examples

The matrix

\[\begin{bmatrix} 1 & 2 & 3\\ 2 & 4 & 5\\ 3 & 5 & 6 \end{bmatrix}\]

and to $(1, 2\sqrt{2}, 3\sqrt{2}, 4, 5\sqrt{2}, 6)$ for PositiveSemidefiniteConeScaled.

source

MathProgBase.Integers — Type.

Integers(dim)

The set of integers $\mathbb{Z}^{dim}$.

source

MathProgBase.Binaries — Type.

Binaries(dim)

The set of binary vectors $\{ 0, 1 \}^{dim}$.

source

MathProgBase.SOS1 — Type.

SOS1(weights)

The set corresponding to the special ordered set (SOS) constraint of type 1. Of the variables in the set, at most one can be nonzero. The $weights$ induce an ordering of the variables; as such, they should be unique values. The $k$-th element in the set corresponds to the $k$-th weight in $weights$. See here for a description of SOS constraints and their potential uses.

source

MathProgBase.SOS2 — Type.

SOS2(weights)

The set corresponding to the special ordered set (SOS) constraint of type 2. Of the variables in the set, at most two can be nonzero, and if two are nonzero, they must be adjacent in the ordering of the set. The $weights$ induce an ordering of the variables; as such, they should be unique values. The $k$-th element in the set corresponds to the $k$-th weight in $weights$. See here for a description of SOS constraints and their potential uses.

source

Attributes

These are used to get and set properties of the model or solver.

MathProgBase.AbstractSolverOrModelAttribute — Type.

AbstractSolverOrModelAttribute

Abstract supertype for attribute objects that can be used to set or get attributes (properties) of the model or solver.

source

MathProgBase.AbstractVariableAttribute — Type.

AbstractVariableAttribute

Abstract supertype for attribute objects that can be used to set or get attributes (properties) of variables in the model.

source

MathProgBase.AbstractConstraintAttribute — Type.

AbstractConstraintAttribute

Abstract supertype for attribute objects that can be used to set or get attributes (properties) of constraints in the model.

source

MathProgBase.cangetattribute — Function.

cangetattribute(s::AbstractMathProgSolver, attr::AbstractSolverOrModelAttribute)::Bool

Return a Bool indicating whether it is possible to query attribute attr from the solver s.

cangetattribute(m::AbstractMathProgModel, attr::AbstractVariableAttribute, R::Type{VariableReference})::Bool
cangetattribute(m::AbstractMathProgModel, attr::AbstractConstraintAttribute, R::Type{VariablewiseConstraintReference{T})::Bool
cangetattribute(m::AbstractMathProgModel, attr::AbstractConstraintAttribute, R::Type{AffineConstraintReference{T})::Bool
cangetattribute(m::AbstractMathProgModel, attr::AbstractConstraintAttribute, R::Type{QuadraticConstraintReference{T})::Bool

Return a Bool indicating whether the model m currently has a value for the attributed specified by attribute type attr applied to the reference type R.

Examples

```julia cangetattribute(GurobiSolver(), SupportsAffineConstraint{Zero}()) cangetattribute(m, ObjectiveValue()) cangetattribute(m, VariablePrimalStart(), varref) cangetattribute(m, ConstraintPrimal(), conref)

source

MathProgBase.getattribute — Function.

getattribute(s::AbstractMathProgSolver, attr::AbstractSolverOrModelAttribute)

Return an attribute attr of the solver s.

getattribute(m::AbstractMathProgModel, attr::AbstractSolverOrModelAttribute)

Return an attribute attr of the model m.

getattribute(m::AbstractMathProgModel, attr::AbstractVariableAttribute, v::VariableReference)

Return an attribute attr of the variable v in model m.

getattribute(m::AbstractMathProgModel, attr::AbstractVariableAttribute, v::Vector{VariableReference})

Return a vector of attributes corresponding to each variable in the collection v in the model m.

getattribute(m::AbstractMathProgModel, attr::AbstractConstraintAttribute, c::ConstraintReference)

Return an attribute attr of the constraint c in model m.

getattribute(m::AbstractMathProgModel, attr::AbstractConstraintAttribute, c::Vector{VariablewiseConstraintReference{T}})
getattribute(m::AbstractMathProgModel, attr::AbstractConstraintAttribute, c::Vector{AffineConstraintReference{T}})
getattribute(m::AbstractMathProgModel, attr::AbstractConstraintAttribute, c::Vector{QuadraticConstraintReference{T}})

Return a vector of attributes corresponding to each constraint in the collection c in the model m.

Examples

getattribute(m, ObjectiveValue())
getattribute(m, VariableResult(), ref)
getattribute(m, VariableResult(5), [ref1,ref2])
getattribute(m, OtherAttribute("something specific to cplex"))

source

MathProgBase.getattribute! — Function.

getattribute!(output, m::AbstractMathProgModel, args...)

An in-place version of getattribute. The signature matches that of getattribute except that the the result is placed in the vector output.

source

MathProgBase.cansetattribute — Function.

cansetattribute(s::AbstractMathProgSolver, attr::AbstractSolverOrModelAttribute)::Bool

Return a Bool indicating whether it is possible to set attribute attr in the solver s.

cansetattribute(m::AbstractMathProgModel, attr::AbstractVariableAttribute, R::Type{VariableReference})::Bool
cansetattribute(m::AbstractMathProgModel, attr::AbstractConstraintAttribute, R::Type{VariablewiseConstraintReference{T})::Bool
cangetattribute(m::AbstractMathProgModel, attr::AbstractConstraintAttribute, R::Type{AffineConstraintReference{T})::Bool
cangetattribute(m::AbstractMathProgModel, attr::AbstractConstraintAttribute, R::Type{QuadraticConstraintReference{T})::Bool

Return a Bool indicating whether it is possible to set attribute attr applied to the reference type R in the model m.

Examples

```julia cansetattribute(GurobiSolver(), SupportsAffineConstraint{Zero}()) cansetattribute(m, ObjectiveValue()) cansetattribute(m, VariablePrimalStart(), VariableReference) cansetattribute(m, ConstraintPrimal(), AffineConstraintReference{NonNegative})

source

MathProgBase.setattribute! — Function.

setattribute!(s::AbstractMathProgSolver, attr::AbstractSolverOrModelAttribute, value)

Assign value to the attribute attr of the solver s.

setattribute!(m::AbstractMathProgModel, attr::AbstractSolverOrModelAttribute, value)

Assign value to the attribute attr of the model m.

setattribute!(m::AbstractMathProgModel, attr::AbstractVariableAttribute, v::VariableReference, value)

Assign value to the attribute attr of variable v in model m.

setattribute!(m::AbstractMathProgModel, attr::AbstractVariableAttribute, v::Vector{VariableReference}, vector_of_values)

Assign a value respectively to the attribute attr of each variable in the collection v in model m.

setattribute!(m::AbstractMathProgModel, attr::AbstractConstraintAttribute, c::ConstraintReference, value)

Assign a value to the attribute attr of constraint c in model m.

setattribute!(m::AbstractMathProgModel, attr::AbstractConstraintAttribute, c::Vector{VariablewiseConstraintReference{T}})
setattribute!(m::AbstractMathProgModel, attr::AbstractConstraintAttribute, c::Vector{AffineConstraintReference{T}})
setattribute!(m::AbstractMathProgModel, attr::AbstractConstraintAttribute, c::Vector{QuadraticConstraintReference{T}})

Assign a value respectively to the attribute attr of each constraint in the collection c in model m.

source

Solver or Model Attributes

These attribute apply to solver or model objects, as noted.

MathProgBase.ReturnsDuals — Type.

ReturnsDuals()

A Bool indicating if the solver should be expected to return dual solutions when appropriate. A solver attribute.

source

MathProgBase.SupportsAddConstraintAfterSolve — Type.

SupportsAddConstraintAfterSolver()

A Bool indicating if the solver supports adding constraints after a solve. If false, then a new model should be constructed instead. A solver attribute.

source

MathProgBase.SupportsDeleteConstraint — Type.

SupportsDeleteConstraint()

A Bool indicating if the solver supports deleting constraints from a model. A solver attribute.

source

MathProgBase.SupportsAddVariableAfterSolver — Type.

SupportsAddVariableAfterSolve()

A Bool indicating if the solver supports adding variables after a solve. In the context of linear programming, this is known as column generation. A solver attribute.

source

MathProgBase.SupportsQuadraticObjective — Type.

SupportsQuadraticObjective()

A Bool indicating if the solver supports quadratic objectives. A solver attribute.

source

MathProgBase.SupportsConicThroughQuadratic — Type.

SupportsConicThroughQuadratic()

A Bool indicating if the solver interprets certain quadratic constraints as second-order cone constraints. A solver attribute.

source

MathProgBase.ObjectiveValue — Type.

ObjectiveValue(resultidx::Int=1, objectiveindex::Int=1)

The objective value of the resultindex'th primal result of the objectiveindex'th objective. A model attribute.

Both resultindex and objectiveindex default to 1.

source

MathProgBase.ObjectiveBound — Type.

ObjectiveBound()

The best known bound on the optimal objective value. A model attribute.

source

MathProgBase.RelativeGap — Type.

RelativeGap()

The final relative optimality gap as optimization terminated. That is, $\frac{|b-f|}{|f|}$, where $b$ is the best bound and $f$ is the best feasible objective value. A model attribute.

source

MathProgBase.SolveTime — Type.

SolveTime()

The total elapsed solution time (in seconds) as reported by the solver. A model attribute.

source

MathProgBase.Sense — Type.

Sense()

The optimization sense of the model, an OptimizationSense with value MinSense or MaxSense. A model attribute.

source

MathProgBase.SimplexIterations — Type.

SimplexIterations()

The cumulative number of simplex iterations during the optimization process. In particular, for a MIP the total simplex iterations for all nodes. A model attribute.

source

MathProgBase.BarrierIterations — Type.

BarrierIterations()

The cumulative number of barrier iterations during the optimization process. A model attribute.

source

MathProgBase.NodeCount — Type.

NodeCount()

The total number of branch-and-bound nodes explored. A model attribute.

source

MathProgBase.RawSolver — Type.

RawSolver()

An object that may be used to access a solver-specific API for this model. A model attribute.

source

MathProgBase.ResultCount — Type.

ResultCount()

The number of results available. A model attribute.

source

MathProgBase.NumberOfVariables — Type.

NumberOfVariables()

The number of variables in the model. A model attribute.

source

MathProgBase.NumberOfVariablewiseConstraints — Type.

NumberOfVariablewiseConstraints{T}()

The number of variablewise constraints of type T in the model. A model attribute.

source

MathProgBase.NumberOfAffineConstraints — Type.

NumberOfAffineConstraints{T}()

The number of affine constraints of type T in the model. A model attribute.

source

MathProgBase.NumberOfQuadraticConstraints — Type.

NumberOfQuadraticConstraints{T}()

The number of quadratic constraints of type T in the model. A model attribute.

source

MathProgBase.SupportsVariablewiseConstraint — Type.

SupportsVariablewiseConstraint{T}()

A Bool indicating whether the solver or model supports a variablewise constraint in the set $S$ which is a set of type T. A solver and model attribute.

source

MathProgBase.SupportsAffineConstraint — Type.

SupportsAffineConstraint{T}()

A Bool indicating whether the solver or model supports a constraint of of the form "affine expression" in $S$ where $S$ is a set of type T. A solver and model attribute.

source

MathProgBase.SupportsQuadraticConstraint — Type.

SupportsQuadraticConstraint{T}()

A Bool indicating whether the solver or model supports a constraint of of the form "quadratic expression" in $S$ where $S$ is a set of type T. A solver and model attribute.

source

MathProgBase.TerminationStatus — Type.

TerminationStatus()

A TerminationStatusCode explaining why the solver stopped. A model attribute.

source

MathProgBase.PrimalStatus — Type.

PrimalStatus(N)
PrimalStatus()

The ResultStatusCode of the primal result N. If N is omitted, it defaults to 1. A model attribute.

source

MathProgBase.DualStatus — Type.

DualStatus(N)
DualStatus()

The ResultStatusCode of the dual result N. If N is omitted, it defaults to 1. A model attribute.

source

Variable Attributes

These attributes are associated with variables. Calls to getattribute and setattribute! should include as an argument a single VariableReference or a vector of VariableReference objects. By convention, the name of a variable attributes begins with Variable.

MathProgBase.VariablePrimalStart — Type.

VariablePrimalStart()

An initial assignment of the variables that the solver may use to warm-start the solve.

source

MathProgBase.VariableLowerBoundDualStart — Type.

VariableLowerBoundDualStart()

An initial assignment of the variable lower-bound duals that the solver may use to warm-start the solve.

source

MathProgBase.VariableUpperBoundDualStart — Type.

VariableUpperBoundDualStart()

An initial assignment of the variable upper-bound duals that the solver may use to warm-start the solve.

source

MathProgBase.VariableLowerBound — Type.

VariableLowerBound()

Lower-bound constraints on variables. -Inf is valid as no bound.

source

MathProgBase.VariableUpperBound — Type.

VariableUpperBound()

Upper-bound constraints for the variables. Inf is valid as no bound.

source

MathProgBase.VariablePrimal — Type.

VariablePrimal(N)
VariablePrimal()

The assignment to the primal variables in result N. If N is omitted, it is 1 by default.

source

MathProgBase.VariableLowerBoundDual — Type.

VariableLowerBoundDual(N)

The assignment to the duals on the variable lower bounds in result N. If N is omitted, it is interpreted as 1.

source

MathProgBase.VariableUpperBoundDual — Type.

VariableUpperBoundDual(N)

The assignment to the duals on the variable upper bounds in result N. If N is omitted, it is interpreted as 1.

source

MathProgBase.VariableBasisStatus — Type.

VariableBasisStatus()

Returns the BasisStatusCode of a given variable, with respect to an available optimal solution basis.

source

Constraint Attributes

These attributes are associated with constraints. Calls to getattribute and setattribute! should include as an argument a single ConstraintReference or a vector of ConstriaintReference{T} objects. By convention, the name of a variable attributes begins with Variable.

MathProgBase.ConstraintPrimalStart — Type.

ConstraintPrimalStart()

An initial assignment of the constraint primal values that the solver may use to warm-start the solve.

source

MathProgBase.ConstraintDualStart — Type.

ConstraintDualStart()

An initial assignment of the constriant duals that the solver may use to warm-start the solve.

source

MathProgBase.ConstraintPrimal — Type.

ConstraintPrimal(N)
ConstraintPrimal()

The assignment to the constraint primal values in result N. If N is omitted, it is 1 by default.

source

MathProgBase.ConstraintDual — Type.

ConstraintDual(N)
ConstraintDual()

The assignment to the constraint dual values in result N. If N is omitted, it is 1 by default.

source

MathProgBase.ConstraintBasisStatus — Type.

ConstraintBasisStatus()

Returns the BasisStatusCode of a given constraint, with respect to an available optimal solution basis.

source

Status Codes

Termination Status

The TerminationStatus attribute is meant to explain the reason why the solver stopped executing. The value of the attribute is of type TerminationStatusCode.

MathProgBase.TerminationStatusCode — Type.

TerminationStatusCode

An Enum of possible values for the TerminationStatus attribute. This attribute is meant to explain the reason why the solver stopped executing.

OK

These are generally OK statuses.

Success: the algorithm ran successfully and has a result. This includes cases where the algorithm converges to an infeasible point (NLP) or converges to a solution of a homogeneous self-dual problem and has a certificate of primal/dual infeasibility.

AlmostSuccess: the algorithm almost ran successfully (e.g., to relaxed convergence tolerances) and has a result.

InfeasibleNoResult: the algorithm stopped because it decided that the problem is infeasible but does not have a result to return.

UnboundedNoResult: the algorithm stopped because it decided that the problem is unbounded but does not have a result to return.

InfeasibleOrUnbounded: the algorithm stopped because it decided that the problem is infeasible or unbounded; no result is available. This occasionally happens during MIP presolve.

Limits

The solver stopped because of some user-defined limit. To be documented: IterationLimit, TimeLimit, NodeLimit, SolutionLimit, MemoryLimit, ObjectiveLimit, NormLimit, OtherLimit.

Problematic

This group of statuses means that something unexpected or problematic happened.

SlowProgress: the algorithm stopped because it was unable to continue making progress towards the solution. AlmostSuccess should be used if there is additional information that relaxed convergence tolerances are satisfied.

To be documented: NumericalError, InvalidModel, InvalidOption, Interrupted, OtherError.

source

Result Status

The PrimalStatus and DualStatus attributes are meant to explain how to interpret the result returned by the solver. The value of the attributes are of type ResultStatusCode.

MathProgBase.ResultStatusCode — Type.

ResultStatusCode

An Enum of possible values for the PrimalStatus and DualStatus attributes. The values indicate how to interpret the result vector.

FeasiblePoint

NearlyFeasiblePoint

InfeasiblePoint

InfeasibilityCertificate

NearlyInfeasibilityCertificate

ReductionCertificate

NearlyReductionCertificate

Unknown

Other

source

Basis Status

MathProgBase.BasisStatusCode — Type.

BasisStatusCode

An Enum of possible values for the VariableBasisStatus and ConstraintBasisStatus attribute. This explains the status of a given element with respect to an optimal solution basis. Possible values are: * Basic: element is in the basis. * Nonbasic: element is not in the basis. * NonbasicAtLower: element is not in the basis and is at its lower bound. * NonbasicAtUpper: element is not in the basis and is at its upper bound. * SuperBasic: element is not in the basis but is also not at one of its bounds.

source

Duals

We take the convention that duals on variable lower bounds should be nonnegative, duals on variable upper bounds should be nonpositive, and duals on closed convex cones should belong to the dual cone.

Nonlinear Programming (NLP)

NLP Methods

MathProgBase.loadnlp! — Function.

loadnlp!(m::AbstractNonlinearModel, numVar, numConstr, l, u, lb, ub, sense::OptimizationSense, d::AbstractNLPEvaluator)

Loads the nonlinear programming problem into the model. The parameter numVar is the number of variables in the problem, numConstr is the number of constraints, l contains the variable lower bounds, u contains the variable upper bounds, lb contains the constraint lower bounds, and ub contains the constraint upper bounds. Sense contains the symbol :Max or :Min, indicating the direction of optimization. The final parameter d is an instance of an AbstractNLPEvaluator, described below, which may be queried for evaluating $f$ and $g$ and their corresponding derivatives.

source

MathProgBase.initialize — Function.

initialize(d::AbstractNLPEvaluator, requested_features::Vector{Symbol})

Must be called before any other methods. The vector requested_features lists features requested by the solver. These may include :Grad for gradients of $f$, :Jac for explicit Jacobians of $g$, :JacVec for Jacobian-vector products, :HessVe for Hessian-vector and Hessian-of-Lagrangian-vector products, :Hess for explicit Hessians and Hessian-of-Lagrangians, and :ExprGraph for expression graphs.

source

MathProgBase.features_available — Function.

features_available(d::AbstractNLPEvaluator)

Returns the subset of features available for this problem instance, as a list of symbols in the same format as in initialize.

source

MathProgBase.eval_f — Function.

eval_f(d::AbstractNLPEvaluator, x)

Evaluate $f(x)$, returning a scalar value.

source

MathProgBase.eval_grad_f — Function.

eval_grad_f(d::AbstractNLPEvaluator, g, x)

Evaluate $\nabla f(x)$ as a dense vector, storing the result in the vector g which must be of the appropriate size.

source

MathProgBase.jac_structure — Function.

jac_structure(d::AbstractNLPEvaluator)

Returns the sparsity structure of the Jacobian matrix, $J_g(x) = \left[ \begin{array}{c} \nabla g_1(x) \\ \nabla g_2(x) \\ \vdots \\ \nabla g_m(x) \end{array} \right]$ where $g_i$ is the $i\text{th}$ component of $g$. The sparsity structure is assumed to be independent of the point $x$. Returns a tuple $(I,J)$ where $I$ contains the row indices and $J$ contains the column indices of each structurally nonzero element. These indices are not required to be sorted and can contain duplicates, in which case the solver should combine the corresponding elements by adding them together.

source

MathProgBase.hesslag_structure — Function.

hesslag_structure(d::AbstractNLPEvaluator)

Returns the sparsity structure of the Hessian-of-the-Lagrangian matrix $\nabla^2 f + \sum_{i=1}^m \nabla^2 g_i$ as a tuple $(I,J)$ where $I$ contains the row indices and $J$ contains the column indices of each structurally nonzero element. These indices are not required to be sorted and can contain duplicates, in which case the solver should combine the corresponding elements by adding them together. Any mix of lower and upper-triangular indices is valid. Elements $(i,j)$ and $(j,i)$, if both present, should be treated as duplicates.

source

MathProgBase.eval_jac_g — Function.

eval_jac_g(d::AbstractNLPEvaluator, J, x)

Evaluates the sparse Jacobian matrix $J_g(x) = \left[ \begin{array}{c} \nabla g_1(x) \\ \nabla g_2(x) \\ \vdots \\ \nabla g_m(x) \end{array} \right]$. The result is stored in the vector J in the same order as the indices returned by jac_structure.

source

MathProgBase.eval_jac_prod — Function.

eval_jac_prod(d::AbstractNLPEvaluator, y, x, w)

Computes the Jacobian-vector product $J_g(x)w$, storing the result in the vector y.

source

MathProgBase.eval_jac_prod_t — Function.

eval_jac_prod_t(d::AbstractNLPEvaluator, y, x, w)

Computes the Jacobian-transpose-vector product $J_g(x)^T w$, storing the result in the vector y.

source

MathProgBase.eval_hesslag_prod — Function.

eval_hesslag_prod(d::AbstractNLPEvaluator, h, x, v, σ, μ)

Given scalar weight $σ$ and vector of constraint weights $μ$, computes the Hessian-of-the-Lagrangian-vector product $\left( \sigma \nabla^2 f(x) + \sum_{i=1}^m \mu_i \nabla^2 g_i(x) \right)v$, storing the result in the vector $h$.

source

MathProgBase.eval_hesslag — Function.

eval_hesslag(d::AbstractNLPEvaluator, H, x, σ, μ)

Given scalar weight σ and vector of constraint weights μ, computes the sparse Hessian-of-the-Lagrangian matrix $\sigma \nabla^2 f(x) + \sum_{i=1}^m \mu_i \nabla^2 g_i(x)$, storing the result in the vector H in the same order as the indices returned by hesslag_structure.

source

MathProgBase.isobjlinear — Method.

isobjlinear(::AbstractNLPEvaluator)

true if the objective function is known to be linear, false otherwise.

source

MathProgBase.isobjquadratic — Method.

isobjquadratic(::AbstractNLPEvaluator)

true if the objective function is known to be quadratic (convex or nonconvex), false otherwise.

source

MathProgBase.isconstrlinear — Method.

isconstrlinear(::AbstractNLPEvaluator, i::Integer)

true if the $i^{\text{th}}$ constraint is known to be linear, false otherwise.

source

MathProgBase.obj_expr — Function.

obj_expr(d::AbstractNLPEvaluator)

Returns an expression graph for the objective function as a standard Julia Expr object. All sums and products are flattened out as simple Expr(:+,...) and Expr(:*,...) objects. The symbol x is used as a placeholder for the vector of decision variables. No other undefined symbols are permitted; coefficients are embedded as explicit values. For example, the expression $x_1+\sin(x_2/\exp(x_3))$ would be represented as the Julia object :(x[1] + sin(x[2]/exp(x[3]))). See the Julia manual for more information on the structure of Expr objects. There are currently no restrictions on recognized functions; typically these will be built-in Julia functions like ^, exp, log, cos, tan, sqrt, etc., but modeling interfaces may choose to extend these basic functions.

source

MathProgBase.constr_expr — Function.

constr_expr(d::AbstractNLPEvaluator, i)

Returns an expression graph for the $i^{\text{th}}$ constraint in the same format as described above. The head of the expression is $:comparison$, indicating the sense of the constraint. The right-hand side of the comparison must be a constant; that is, :(x[1]^3 <= 1) is allowed, while :(1 <= x[1]^3) is not valid. Double-sided constraints are allowed, in which case both the lower bound and upper bounds should be constants; for example, :(-1 <= cos(x[1]) + sin(x[2]) <= 1) is valid.

source

NLP Attributes

MathProgBase.ConstraintNLPDual — Type.

ConstraintNLPDual(N)
ConstraintNLPDual()

The assignment to the NLP constraint dual values in result N. If N is omitted, it is 1 by default.

source

MathProgBase.ConstraintNLPDualStart — Type.

ConstraintNLPDualStart()

An initial assignment of the NLP constriant duals that the solver may use to warm-start the solve.

source